rnd: Fix compiler warning
Fix the following compiler warning that occurs when building with "-W
-Wall -Wextra":

rnd.c: In function ‘randombytes_strong’:
rnd.c:50:10: warning: comparison between signed and unsigned integer expressions [-Wsign-compare]

+++ rnd.c
@@ -47,7 +47,7 @@ static void randombytes_strong(unsigned char *x, size_t xlen)
 	fds = open_or_die(HIG_ENTROPY_SOURCE, O_RDONLY);
 
 	ret = read_exact(fds, x, xlen, 0);
-	if (ret != xlen)
+	if (ret != (int) xlen)
 		panic("Error reading from entropy source!\n");
 
 	close(fds);
